docs: update button docs
